fix naming tests
