+++ internal/introspect/introspect_test.go
@@ -15,7 +15,6 @@
 package introspect_test
 
 import (
-	"context"
 	"fmt"
 	"reflect"
 	"testing"
@@ -24,7 +23,6 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/pulumi/pulumi/sdk/v3/go/common/tokens"
-	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
 
 	"github.com/pulumi/pulumi-go-provider/infer"
 	"github.com/pulumi/pulumi-go-provider/internal/introspect"
@@ -45,44 +43,6 @@ func (m *MyStruct) Annotate(a infer.Annotator) {
 	a.SetToken("myMod", "MyToken")
 	a.Deprecate(&m, "This resource is deprecated.")
 	a.AddAlias("myMod", "MyAlias")
-}
-
-type (
-	MyResource       struct{}
-	MyResourceArgs   struct{}
-	MyResourceOutput struct{}
-)
-
-func (MyResource) Create(
-	ctx context.Context,
-	_ infer.CreateRequest[MyResourceArgs],
-) (infer.CreateResponse[MyResourceOutput], error) {
-	return infer.CreateResponse[MyResourceOutput]{}, nil
-}
-
-type MyComponentArgs struct{}
-
-type MyComponentOutput struct {
-	pulumi.ResourceState
-}
-
-func (MyComponentArgs) Create(
-	ctx context.Context,
-	_ infer.CreateRequest[MyComponentArgs],
-) (infer.CreateResponse[MyComponentOutput], error) {
-	return infer.CreateResponse[MyComponentOutput]{}, nil
-}
-
-type MyComponent struct{}
-
-func (MyComponent) Construct(
-	ctx *pulumi.Context,
-	name string,
-	typ string,
-	args MyStruct,
-	opts pulumi.ResourceOption,
-) (*MyComponentOutput, error) {
-	return &MyComponentOutput{}, nil
 }
 
 func TestParseTag(t *testing.T) {
@@ -240,26 +200,4 @@ func TestAllFieldsMiss(t *testing.T) {
 	_, ok, err := fm.TargetStructFields(&s.Fizz)
 	require.False(t, ok)
 	assert.NoError(t, err)
-}
-
-func TestGetToken(t *testing.T) {
-	t.Parallel()
-
-	t.Run("component", func(t *testing.T) {
-		t.Parallel()
-
-		component := infer.Component(&MyComponent{})
-		tok, err := introspect.GetToken(tokens.NewPackageToken("pkg"), reflect.TypeOf(component))
-		require.NoError(t, err)
-		assert.Equal(t, tokens.TypeName("MyComponentOutput"), tok.Name())
-	})
-
-	t.Run("resource", func(t *testing.T) {
-		t.Parallel()
-
-		resource := infer.Resource(&MyResource{})
-		tok, err := introspect.GetToken(tokens.NewPackageToken("pkg"), reflect.TypeOf(resource))
-		require.NoError(t, err)
-		assert.Equal(t, tokens.TypeName("MyResourceOutput"), tok.Name())
-	})
 }
